VP-429: extend the cart service confirmation popups were moved into cart service new methods

+++ ng-app/src/app/services/cart.service.ts
@@ -1,3 +1,4 @@
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { Injectable } from '@angular/core';
 import { ActiveOrderService } from '@api/active-order.service';
 import { AlertsService } from '@modules/alerts/alerts.service';
@@ -5,6 +6,7 @@ import { ConfirmService } from '@modules/confirm-modal/confirm-modal-service';
 import { Subject } from 'rxjs';
 import { ILineItem, ICart } from '@models/dto/icart';
 import { shareReplay, debounceTime, distinctUntilChanged } from 'rxjs/operators';
+import { CheckoutModalComponent } from '@components/active-order/checkout-modal/checkout-modal.component';
 
 @Injectable({
   providedIn: 'root'
@@ -16,7 +18,8 @@ export class CartService {
 
   constructor(private readonly activeOrderService: ActiveOrderService,
               private confirmService: ConfirmService,
-              private alertsService: AlertsService) {
+              private alertsService: AlertsService,
+              private modalService: NgbModal) {
 
     this.productQuantityChanging$.pipe(
       shareReplay(),
@@ -40,9 +43,13 @@ export class CartService {
     return product != null ? true : false;
   }
 
-  public inCartQuantity(productId: string): number {
-    if (this.isInCart(productId)) {
-      return this.findProductInCart(this.cart, productId).quantity;
+  public inCartQuantity(productId?: string): number {
+    if (productId != null) {
+      if (this.isInCart(productId)) {
+        return this.findProductInCart(this.cart, productId).quantity;
+      }
+    } else {
+      return this.cart.itemsCount;
     }
 
     return 0;
@@ -75,7 +82,7 @@ export class CartService {
     if (product.quantity < 1) {
       // by default = 1
       product.quantity = 1;
-      const dialogResult = await this.showApproveDeletionConfirmation(product.quantity);
+      const dialogResult = await this.showRemoveConfirmation();
       if (dialogResult === true) {
         this.activeOrderService.removeItem(product.id).subscribe();
         return;
@@ -91,15 +98,45 @@ export class CartService {
     }
   }
 
-  public remove(lineItemId: string) {
-    this.activeOrderService.removeItem(lineItemId).subscribe();
+  public async remove(lineItemId: string) {
+    if (await this.showRemoveConfirmation()) {
+      this.activeOrderService.removeItem(lineItemId).subscribe();
+    }
   }
 
-  private showApproveDeletionConfirmation(newQuantity: number): Promise<boolean> {
+  public async removeAll() {
+    if (await this.showRemoveAllConfirmation()) {
+      this.activeOrderService.clearAllItems().subscribe();
+    }
+  }
+
+  public async checkout() {
+    const modalRef = this.modalService.open(CheckoutModalComponent, {
+      centered: true,
+      backdrop: 'static',
+      size: 'lg'
+    });
+    modalRef.componentInstance.cart = this.cart;
+    await modalRef.result.then(() => {
+      this.activeOrderService.createOrder().subscribe(() => this.alertsService.success(`Order is created successfully!`));
+    }, () => { });
+  }
+
+  private showRemoveConfirmation(): Promise<boolean> {
     const confirmOptions = {
       title: 'Line item removing',
       backdrop: 'static',
       message: 'Are you sure you want to remove this line item from the active order?'
+    };
+
+    return this.confirmService.confirm(confirmOptions).then(() => true, () => false);
+  }
+
+  private showRemoveAllConfirmation() {
+    const confirmOptions = {
+      title: 'Active order cleaning',
+      backdrop: 'static',
+      message: 'Are you sure you want to clear the active order?'
     };
 
     return this.confirmService.confirm(confirmOptions).then(() => true, () => false);
